+++ common/src/main/java/io/github/mortuusars/thief/Thief.java
@@ -97,6 +97,13 @@ public class Thief {
             public static final TagKey<EntityType<?>> KILLING_PROTECTED_HEAVY =
                     TagKey.create(net.minecraft.core.registries.Registries.ENTITY_TYPE, identifier("killing_protected/heavy"));
 
+            public static final TagKey<EntityType<?>> PICKUP_PROTECTED_LIGHT =
+                  TagKey.create(net.minecraft.core.registries.Registries.ENTITY_TYPE, resource("pickup_protected/light"));
+            public static final TagKey<EntityType<?>> PICKUP_PROTECTED_MEDIUM =
+                  TagKey.create(net.minecraft.core.registries.Registries.ENTITY_TYPE, resource("pickup_protected/medium"));
+            public static final TagKey<EntityType<?>> PICKUP_PROTECTED_HEAVY =
+                  TagKey.create(net.minecraft.core.registries.Registries.ENTITY_TYPE, resource("pickup_protected/heavy"));
+
             public static final TagKey<EntityType<?>> WITNESSES =
                     TagKey.create(net.minecraft.core.registries.Registries.ENTITY_TYPE, identifier("witnesses"));