+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## UNRELEASED
 - Looting an Item Frame is now a crime.
+- Picking up entities with CarryOn mod is now a crime.
 
 ## 1.2.4 - 2026-06-12
 - [NeoForge] Support for 26.2 NeoForge.